fr-steganography-1/code/requirements.txt
numpy==2.2.3
Pillow==11.1.0
